samwriter as its own module
