+++ src/bin/test_bam_decompress.rs
@@ -1,7 +1,8 @@
 extern crate rust_htslib;
-use self::rust_htslib::bam::SAMWriter;
+use self::rust_htslib::sam::SAMWriter;
 
 use std::io::Write;
+
 
 pub fn main(){
   let example_bam = "./test/bam2sam_test.bam";
